@@ -28,86 +28,89 @@
     "update": "nx migrate latest",
     "dep-graph": "nx dep-graph",
     "help": "nx help",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "workspace-generator": "nx workspace-generator"
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^13.2.3",
-    "@angular/animations": "13.2.2",
-    "@angular/cdk": "13.2.2",
-    "@angular/common": "13.2.2",
-    "@angular/compiler": "13.2.2",
-    "@angular/core": "13.2.2",
+    "@angular-devkit/schematics": "15.1.6",
+    "@angular/animations": "15.1.5",
+    "@angular/cdk": "15.1.5",
+    "@angular/common": "15.1.5",
+    "@angular/compiler": "15.1.5",
+    "@angular/core": "15.1.5",
     "@angular/fire": "^7.5.0",
-    "@angular/flex-layout": "13.0.0-beta.38",
-    "@angular/forms": "13.2.2",
-    "@angular/google-maps": "13.2.2",
-    "@angular/material": "13.2.2",
-    "@angular/platform-browser": "13.2.2",
-    "@angular/platform-browser-dynamic": "13.2.2",
-    "@angular/router": "13.2.2",
-    "@angular/service-worker": "13.2.2",
-    "@fortawesome/fontawesome-pro": "6.1.0",
-    "@ngrx/component-store": "13.0.2",
-    "@ngrx/effects": "13.0.2",
-    "@ngrx/entity": "13.0.2",
-    "@ngrx/router-store": "13.0.2",
-    "@ngrx/store": "13.0.2",
-    "@ngrx/store-devtools": "13.0.2",
-    "@nrwl/angular": "13.8.1",
-    "@swimlane/ngx-charts": "^20.0.1",
+    "@angular/flex-layout": "^15.0.0-beta.42",
+    "@angular/forms": "15.1.5",
+    "@angular/google-maps": "^15.1.5",
+    "@angular/material": "15.1.5",
+    "@angular/platform-browser": "15.1.5",
+    "@angular/platform-browser-dynamic": "15.1.5",
+    "@angular/router": "15.1.5",
+    "@angular/service-worker": "15.1.5",
+    "@fortawesome/fontawesome-pro": "^6.3.0",
+    "@ngrx/component-store": "^15.3.0",
+    "@ngrx/effects": "^15.3.0",
+    "@ngrx/entity": "^15.3.0",
+    "@ngrx/router-store": "^15.3.0",
+    "@ngrx/store": "^15.3.0",
+    "@ngrx/store-devtools": "^15.3.0",
+    "@nrwl/angular": "15.7.2",
+    "@swimlane/ngx-charts": "^20.1.2",
     "axios": "^0.19.2",
     "core-js": "^2.5.4",
     "firebase": "^9.17.1",
     "firebase-admin": "^11.5.0",
-    "luxon": "^2.3.0",
+    "luxon": "^3.2.1",
     "mysql": "^2.18.1",
-    "rxjs": "7.5.4",
+    "rxjs": "~7.5.0",
     "tslib": "^2.0.0",
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.1302.3",
-    "@angular-devkit/build-angular": "13.2.3",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "13.2.3",
-    "@angular/compiler-cli": "13.2.2",
-    "@angular/language-service": "13.2.2",
-    "@ngrx/schematics": "13.0.2",
-    "@ngrx/store-devtools": "13.0.2",
-    "@nrwl/cli": "13.8.1",
-    "@nrwl/cypress": "13.8.1",
-    "@nrwl/eslint-plugin-nx": "13.8.1",
-    "@nrwl/jest": "13.8.1",
-    "@nrwl/linter": "13.8.1",
-    "@nrwl/node": "13.8.1",
-    "@nrwl/tao": "13.8.1",
-    "@nrwl/workspace": "13.8.1",
-    "@types/jest": "27.0.2",
-    "@types/luxon": "^2.0.9",
+    "@angular-devkit/architect": "0.1501.6",
+    "@angular-devkit/build-angular": "15.1.6",
+    "@angular-devkit/core": "~15.1.0",
+    "@angular-eslint/eslint-plugin": "15.0.0",
+    "@angular-eslint/eslint-plugin-template": "15.0.0",
+    "@angular-eslint/template-parser": "15.0.0",
+    "@angular/cli": "~15.1.0",
+    "@angular/compiler-cli": "15.1.5",
+    "@angular/language-service": "15.1.5",
+    "@ngrx/schematics": "^15.3.0",
+    "@ngrx/store-devtools": "^15.3.0",
+    "@nrwl/cli": "15.7.2",
+    "@nrwl/cypress": "15.7.2",
+    "@nrwl/eslint-plugin-nx": "15.7.2",
+    "@nrwl/jest": "15.7.2",
+    "@nrwl/linter": "15.7.2",
+    "@nrwl/node": "15.7.2",
+    "@nrwl/workspace": "15.7.2",
+    "@schematics/angular": "~15.1.0",
+    "@types/jest": "28.1.8",
+    "@types/luxon": "^3.2.0",
     "@types/mysql": "^2.15.15",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "5.10.2",
-    "@typescript-eslint/parser": "5.10.2",
+    "@types/node": "18.7.1",
+    "@typescript-eslint/eslint-plugin": "5.52.0",
+    "@typescript-eslint/parser": "5.52.0",
     "codelyzer": "6.0.1",
     "cypress": "^3.8.2",
     "dotenv": "10.0.0",
-    "eslint": "8.7.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.11.2",
     "firebase-tools": "^11.23.1",
     "fuzzy": "^0.1.3",
     "inquirer": "^6.2.2",
     "inquirer-autocomplete-prompt": "^1.2.0",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
-    "prettier": "2.5.1",
-    "ts-jest": "27.0.5",
-    "ts-node": "9.1.1",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.3",
+    "jest-preset-angular": "12.2.3",
+    "nx": "15.7.2",
+    "prettier": "2.8.4",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
     "tslint": "6.1.3",
-    "typescript": "4.5.5"
+    "typescript": "4.8.4"
   }
 }
