@@ -1,4 +1,5 @@
-module.exports = {
+/* eslint-disable */
+export default {
   preset: "../../jest.preset.js",
   transform: {
     "^.+\\.[tj]sx?$": "ts-jest",
