@@ -1,21 +1,21 @@
-module.exports = {
-  displayName: "teams",
+/* eslint-disable */
+export default {
   preset: "../../jest.preset.js",
+  coverageDirectory: "../../coverage/libs/shared",
+
   setupFilesAfterEnv: ["<rootDir>/src/test-setup.ts"],
   globals: {
     "ts-jest": {
-      tsconfig: "<rootDir>/tsconfig.spec.json",
       stringifyContentPathRegex: "\\.(html|svg)$",
+
+      tsconfig: "<rootDir>/tsconfig.spec.json",
     },
   },
-  coverageDirectory: "../../coverage/libs/teams",
-  transform: {
-    "^.+\\.(ts|mjs|js|html)$": "jest-preset-angular",
-  },
-  transformIgnorePatterns: ["node_modules/(?!.*\\.mjs$)"],
+  displayName: "shared",
   snapshotSerializers: [
     "jest-preset-angular/build/serializers/no-ng-attributes",
     "jest-preset-angular/build/serializers/ng-snapshot",
     "jest-preset-angular/build/serializers/html-comment",
   ],
+  transform: { "^.+\\.(ts|js|html)$": "jest-preset-angular" },
 };
