@@ -1,20 +1,22 @@
-module.exports = {
+/* eslint-disable */
+export default {
+  displayName: "teams",
   preset: "../../jest.preset.js",
-  coverageDirectory: "../../coverage/libs/race",
-
   setupFilesAfterEnv: ["<rootDir>/src/test-setup.ts"],
   globals: {
     "ts-jest": {
-      stringifyContentPathRegex: "\\.(html|svg)$",
-
       tsconfig: "<rootDir>/tsconfig.spec.json",
+      stringifyContentPathRegex: "\\.(html|svg)$",
     },
   },
-  displayName: "race",
+  coverageDirectory: "../../coverage/libs/teams",
+  transform: {
+    "^.+\\.(ts|mjs|js|html)$": "jest-preset-angular",
+  },
+  transformIgnorePatterns: ["node_modules/(?!.*\\.mjs$)"],
   snapshotSerializers: [
     "jest-preset-angular/build/serializers/no-ng-attributes",
     "jest-preset-angular/build/serializers/ng-snapshot",
     "jest-preset-angular/build/serializers/html-comment",
   ],
-  transform: { "^.+\\.(ts|js|html)$": "jest-preset-angular" },
 };
