@@ -1,4 +1,4 @@
-const nxPreset = require("@nrwl/jest/preset");
+const nxPreset = require('@nrwl/jest/preset').default;
 module.exports = {
   ...nxPreset,
   testMatch: ["**/+(*.)+(spec|test).+(ts|js)?(x)"],
